pyproject.toml
[tool.poetry]
name = "ioc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11" 
black = "^23.7.0"
sqlalchemy = "^2.0.19"
fastapi = "^0.101.1"
httpx = "^0.24.1"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
psycopg2-binary = "^2.9.7"
dynaconf = "^3.2.1"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
